fix: compression type u8 in header

+++ msg-socket/src/pub/mod.rs
@@ -144,7 +144,7 @@ impl PubMessage {
             seq,
             Bytes::from(self.topic),
             self.payload,
-            self.compression_type,
+            self.compression_type as u8,
         )
     }
